added test background to class diagram
