+++ Application/src/view/visuals/component/ClassDiagramClass.java
@@ -2,6 +2,7 @@ package view.visuals.component;
 
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.image.Image;
+import javafx.scene.paint.Color;
 import view.visuals.Renderable;
 
 public class ClassDiagramClass implements Renderable {
@@ -17,6 +18,24 @@ public class ClassDiagramClass implements Renderable {
     @Override
     public void render(GraphicsContext gc) {
 
+        gc.setFill(Color.TRANSPARENT);
+
+        int x = this.coordinates.getX();
+        int y = this.coordinates.getY();
+
+        int size_ClassDiagramClass = size/2;
+
+        placeGraphicCentered(gc, island, size_ClassDiagramClass,0, 0);
+
+    }
+
+    void placeGraphicCentered(GraphicsContext gc, Image img, int size, int offsetX, int offsetY) {
+        gc.drawImage(
+                img,
+                this.coordinates.getX() -size/2 + offsetX,
+                this.coordinates.getY() -size/2 + offsetY,
+                size,
+                size*(img.getHeight()/img.getWidth()));
     }
 
     @Override
@@ -36,7 +55,8 @@ public class ClassDiagramClass implements Renderable {
 
     @Override
     public String getName() {
-        return null;
+        String[] s = name.split(":");
+        return s[1];
     }
 
     @Override
